Adding themeleaf templates for static pages like index, public, user and admin
