+++ spring-boot4-security-2/src/main/resources/templates/index.html
@@ -1,0 +1,37 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org"
+      xmlns:sec="http://www.thymeleaf.org/extras/spring-security">
+<head>
+    <meta charset="UTF-8">
+    <title>Spring Boot Tutorials</title>
+</head>
+<body>
+    <h1 style="text-align: center;">Spring Boot Tutorial</h1>
+    <h2 style="text-align: center;">Home Page!</h2>
+    <div sec:authorize="isAuthenticated()">
+        <p style="text-align: center;">Username:
+            <span sec:authentication="name" th:text=${#strings.toUpperCase(#authentication.name)}>Anonymous</span>
+        </p>
+    </div>
+    <div sec:authorize="isAnonymous()">
+        <p style="text-align: center;">Username: Anonymous</p>
+    </div>
+    <p style="text-align: center;"><a href='/public'>Public</a></p>
+    <p>&nbsp;</p>
+    <div sec:authorize="!isAuthenticated()">
+        <p style="text-align: center;"><a href='/login'>Login</a></p>
+    </div>
+    <div sec:authorize="isAuthenticated()">
+        <p style="text-align: center;"><a href='/logout'>Logout</a></p>
+    </div>
+    <p>&nbsp;</p>
+    <div sec:authorize="isAuthenticated()">
+        <div sec:authorize="hasRole('ROLE_USER')">
+            <p style="text-align: center;"><a href='/user'>Private for User</a></p>
+        </div>
+        <div sec:authorize="hasRole('ROLE_ADMIN')">
+            <p style="text-align: center;"><a href='/admin'>Private for Admin</a></p>
+        </div>
+    </div>
+</body>
+</html>
