+++ spring-boot4-security-2/src/main/resources/templates/public.html
@@ -1,0 +1,21 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org"
+      xmlns:sec="http://www.thymeleaf.org/extras/spring-security">
+<head>
+    <meta charset="UTF-8">
+    <title>Spring Boot Tutorials</title>
+</head>
+<body>
+    <h1 style="text-align: center;">Spring Boot Tutorial</h1>
+    <h2 style="text-align: center;">Public Page!</h2>
+    <div sec:authorize="isAuthenticated()">
+        <p style="text-align: center;">Username:
+            <span sec:authentication="name" th:text=${#strings.toUpperCase(#authentication.name)}>Anonymous</span>
+        </p>
+    </div>
+    <div sec:authorize="isAnonymous()">
+        <p style="text-align: center;">Username: Anonymous</p>
+    </div>
+    <p style="text-align: center;"><a href='/'>Home</a></p>
+</body>
+</html>
