+++ spring-boot4-security-2/src/main/resources/templates/user.html
@@ -1,0 +1,23 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org"
+      xmlns:sec="http://www.thymeleaf.org/extras/spring-security">
+<head>
+    <meta charset="UTF-8">
+    <title>Spring Boot Tutorials</title>
+</head>
+<body>
+    <h1 style="text-align: center;">Spring Boot Tutorial</h1>
+    <h2 style="text-align: center;">Private Page for User!</h2>
+    <p style="text-align: center;">Username:
+        <span sec:authentication="name" th:text=${#strings.toUpperCase(#authentication.name)}>Anonymous</span>
+    </p>
+    <p style="text-align: center;">Name: <span th:text="${user.name}">Anonymous</span></p>
+    <p style="text-align: center;">E-mail: <span th:text="${user.email}">Anonymous</span></p>
+    <p style="text-align: center;">
+        <span th:if="${user.roles.size()>1}">Roles: </span>
+        <span th:unless="${user.roles.size()>1}">Role: </span>
+        <span th:text="${user.roles}">Anonymous</span>
+    </p>
+    <p style="text-align: center;"><a href='/'>Home</a></p>
+</body>
+</html>
